requirements.txt
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
importlib-resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pkgutil-resolve-name==1.3.10
psycopg==3.2.6
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.1
uritemplate==4.1.1
zipp==3.20.2
